+++ code/test/e2e/loggedin/logger.spec.ts
@@ -34,7 +34,7 @@ test.beforeAll(({}, { config }) => {
   }
 });
 
-test('test', async ({ page }, { config }) => {
+test('testing logger component', async ({ page }, { config }) => {
   const { baseURL } = config.projects[0].use;
 
   await page.routeFromHAR('./test/e2e/loggedin/logger_test.har', { url: '**/api/**' });